backend/requirements.txt
python-docx>=1.1.0
reportlab>=4.0.0
openpyxl>=3.1.0
Flask>=3.0.0
flask-cors>=4.0.0
supabase>=2.3.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
